+++ src/Kanvas/Apps/Models/Roles.php
@@ -4,18 +4,17 @@ declare(strict_types=1);
 
 namespace Kanvas\Apps\Models;
 
+use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Kanvas\Models\BaseModel;
-use Kanvas\Apps\Models\Apps;
 
 /**
- * AppsRoles Class
+ * AppsRoles Class.
  *
  * @property int $apps_id
  * @property string $roles_name
  */
 class Roles extends BaseModel
 {
-
     /**
      * The table associated with the model.
      *
@@ -24,11 +23,11 @@ class Roles extends BaseModel
     protected $table = 'apps_roles';
 
     /**
-     * Apps relationship
+     * Apps relationship.
      *
-     * @return Apps
+     * @return BelongsTo
      */
-    public function app(): Apps
+    public function app() : BelongsTo
     {
         return $this->belongsTo(Apps::class, 'apps_id');
     }
